pyproject.toml
[tool.poetry]
name = "hooks"
version = "0.1.0"
description = ""
authors = ["Вадим Акимов <sfc.coltrane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.8"
djangorestframework = "^3.14"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
redis = "^5.0.1"
psycopg2 = "^2.9.9"
python-keycloak = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
